See 290cc832

+++ tests/conftest.py
@@ -1,7 +1,8 @@
 import pytest
 import os
 
-from app.domain.constants import TEST_DATABASE_FILE
+from app.domain.config import *
+from app.domain.constants import *
 from tests.utils import get_db, setup, clear_db_data
 
 
@@ -18,4 +19,5 @@ def pytest_runtest_teardown(item, nextitem):
 
 @pytest.hookimpl()
 def pytest_sessionfinish(session, exitstatus):
-    os.remove(TEST_DATABASE_FILE)
+    if ENVIRONMENT == "TEST":
+        os.remove(TEST_DATABASE_FILE)
